+++ sdk/include/ila/event/minecraft/server/SendPacketEvent.h
@@ -1,72 +1,117 @@
+#pragma once
 #include "ila/base/Macro.h"
 #include <ll/api/event/Cancellable.h>
+#include <mc/common/SubClientId.h>
 #include <mc/deps/core/utility/optional_ref.h>
 
 // clang-format off
-class LoopbackPacketSender;
+class NetworkSystem;
+class NetworkIdentifier;
 class Packet;
 class ServerPlayer;
 // clang-format on
 
 namespace ila::mc::inline server
 {
-class SendPacketBeforeEvent final : public ll::event::Cancellable<ll::event::Event>
+class ISendPacketBeforeEvent : public ll::event::Cancellable<ll::event::Event>
 {
-protected:
-    LoopbackPacketSender&      mPacketSender;
-    Packet&                    mPacket;
-    bool const&                mIsBroadcast;
-    optional_ref<ServerPlayer> mPlayer;
+private:
+    NetworkSystem&           mNetworkSystem;
+    Packet&                  mPacket;
+    NetworkIdentifier const& mNetworkIdentifier;
+    SubClientId&             mSenderSubId;
 
+public:
+    constexpr explicit ISendPacketBeforeEvent(
+        NetworkSystem&           networkSystem,
+        Packet&                  packet,
+        NetworkIdentifier const& networkIdentifier,
+        SubClientId&             senderSubId
+    )
+        : Cancellable()
+        , mNetworkSystem(networkSystem)
+        , mPacket(packet)
+        , mNetworkIdentifier(networkIdentifier)
+        , mSenderSubId(senderSubId)
+    {
+    }
+
+    ILAPI void serialize(CompoundTag& nbt) const override;
+    ILAPI void deserialize(CompoundTag const& nbt) override;
+
+    ILNDAPI NetworkSystem&           networkSystem() const;
+    ILNDAPI Packet&                  packet() const;
+    ILNDAPI NetworkIdentifier const& networkIdentifier() const;
+    ILNDAPI SubClientId&             senderSubId() const;
+    ILNDAPI optional_ref<ServerPlayer> player() const;
+};
+
+template<std::derived_from<Packet> PacketType = Packet>
+class SendPacketBeforeEvent final : public ISendPacketBeforeEvent
+{
 public:
     constexpr explicit SendPacketBeforeEvent(
-        LoopbackPacketSender&      packetSender,
-        Packet&                    packet,
-        bool const&                isBroadcast,
-        optional_ref<ServerPlayer> player
+        NetworkSystem&           networkSystem,
+        PacketType&              packet,
+        NetworkIdentifier const& networkIdentifier,
+        SubClientId&             senderSubId
     )
-        : mPacketSender(packetSender)
-        , mPacket(packet)
-        , mIsBroadcast(isBroadcast)
-        , mPlayer(player)
+        : ISendPacketBeforeEvent(networkSystem, packet, networkIdentifier, senderSubId)
     {
     }
 
-    ILAPI void serialize(CompoundTag& nbt) const override;
+    PacketType& packet() const { return static_cast<PacketType&>(ISendPacketBeforeEvent::packet()); }
+};
 
-    ILNDAPI LoopbackPacketSender& packetSender() const;
-    ILNDAPI Packet&               packet() const;
-    ILNDAPI bool const&           isBroadcast() const;
-    ILNDAPI optional_ref<ServerPlayer> player() const;
-}; // class SendPacketEvent
-
-class SendPacketAfterEvent final : public ll::event::Event
+class ISendPacketAfterEvent : public ll::event::Event
 {
-protected:
-    LoopbackPacketSender&      mPacketSender;
-    Packet const&              mPacket;
-    bool const&                mIsBroadcast;
-    optional_ref<ServerPlayer> mPlayer;
+private:
+    NetworkSystem&           mNetworkSystem;
+    Packet const&            mPacket;
+    NetworkIdentifier const& mNetworkIdentifier;
+    SubClientId const&       mSenderSubId;
 
 public:
-    constexpr explicit SendPacketAfterEvent(
-        LoopbackPacketSender&      packetSender,
-        Packet const&              packet,
-        bool const&                isBroadcast,
-        optional_ref<ServerPlayer> player
+    constexpr explicit ISendPacketAfterEvent(
+        NetworkSystem&           networkSystem,
+        Packet const&            packet,
+        NetworkIdentifier const& networkIdentifier,
+        SubClientId const&       senderSubId
     )
-        : mPacketSender(packetSender)
+        : Event()
+        , mNetworkSystem(networkSystem)
         , mPacket(packet)
-        , mIsBroadcast(isBroadcast)
-        , mPlayer(player)
+        , mNetworkIdentifier(networkIdentifier)
+        , mSenderSubId(senderSubId)
     {
     }
 
     ILAPI void serialize(CompoundTag& nbt) const override;
 
-    ILNDAPI LoopbackPacketSender& packetSender() const;
-    ILNDAPI Packet const&         packet() const;
-    ILNDAPI bool const&           isBroadcast() const;
+    ILNDAPI NetworkSystem&           networkSystem() const;
+    ILNDAPI Packet const&            packet() const;
+    ILNDAPI NetworkIdentifier const& networkIdentifier() const;
+    ILNDAPI SubClientId const&       senderSubId() const;
     ILNDAPI optional_ref<ServerPlayer> player() const;
-}; // class SendPacketEvent
+};
+
+template<std::derived_from<Packet> PacketType = Packet>
+class SendPacketAfterEvent final : public ISendPacketAfterEvent
+{
+public:
+    constexpr explicit SendPacketAfterEvent(
+        NetworkSystem&           networkSystem,
+        PacketType const&        packet,
+        NetworkIdentifier const& networkIdentifier,
+        SubClientId const&       senderSubId
+    )
+        : ISendPacketAfterEvent(networkSystem, packet, networkIdentifier, senderSubId)
+    {
+    }
+
+    PacketType const& packet() const
+    {
+        return static_cast<PacketType const&>(ISendPacketAfterEvent::packet());
+    }
+};
 } // namespace ila::mc::inline server