+++ sdk/include/ila/event/minecraft/world/actor/player/PlayerAteEvent.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "ila/base/Macro.h"
+#include <ll/api/event/Cancellable.h>
 #include <ll/api/event/player/PlayerEvent.h>
 
 // clang-format off
@@ -8,21 +9,37 @@ class ItemStack;
 
 namespace ila::mc::inline world::inline actor::inline player
 {
-class PlayerAteEvent final : public ll::event::PlayerEvent
+class PlayerAteBeforeEvent final : public ll::event::Cancellable<ll::event::PlayerEvent>
 {
 protected:
     ItemStack& mItem;
 
 public:
-    constexpr explicit PlayerAteEvent(Player& player, ItemStack& item)
-        : PlayerEvent(player)
+    constexpr explicit PlayerAteBeforeEvent(Player& player, ItemStack& item)
+        : Cancellable(player)
         , mItem(item)
     {
     }
 
     ILAPI void serialize(CompoundTag& nbt) const override;
-    ILAPI void deserialize(CompoundTag const& nbt) override;
 
     ILNDAPI ItemStack& item() const;
 };
-} // namespace ila::mc::inline player
+
+class PlayerAteAfterEvent final : public ll::event::PlayerEvent
+{
+protected:
+    int mSlot;
+
+public:
+    constexpr explicit PlayerAteAfterEvent(Player& player, int slot)
+        : PlayerEvent(player)
+        , mSlot(slot)
+    {
+    }
+
+    ILAPI void serialize(CompoundTag& nbt) const override;
+
+    ILNDAPI int slot() const;
+};
+} // namespace ila::mc::inline world::inline actor::inline player