+++ sdk/include/ila/include_all.h
@@ -7,6 +7,7 @@
 #include "ila/event/minecraft/server/ClientLoginEvent.h"
 #include "ila/event/minecraft/server/ReceivePacketEvent.h"
 #include "ila/event/minecraft/server/RegisterCmdEvent.h"
+#include "ila/event/minecraft/server/SendMultiplePacketEvent.h"
 #include "ila/event/minecraft/server/SendPacketEvent.h"
 #include "ila/event/minecraft/server/ServerPongEvent.h"
 #include "ila/event/minecraft/world/actor/ActorChangeDimensionEvent.h"